requirements.txt
environs==9.5.0
python-telegram-bot==13.11
requests==2.27.1
